Refactor experimental controllers under platform-operator/experimental, move generated experimental manifests out of local dir for repo checks
