+++ Makefile
@@ -40,6 +40,8 @@ PARENT_BRANCH ?= origin/master
 GO ?= CGO_ENABLED=0 GO111MODULE=on GOPRIVATE=github.com/verrazzano go
 GO_LDFLAGS ?= -extldflags -static -X main.buildVersion=${BUILDVERSION} -X main.buildDate=${BUILDDATE}
 
+SOURCE_EXCLUSIONS = tests/e2e|tools/psr|platform-operator/experimental
+
 .PHONY: clean
 clean: ## remove coverage and test results
 	find . -name coverage.cov -exec rm {} \;
@@ -101,7 +103,7 @@ precommit-check: check check-tests copyright-check ## run precommit checks witho
 precommit-build:  ## go build the project
 	go build ./...
 
-unit-test-coverage: export COVERAGE_EXCLUSIONS ?= tests/e2e|tools/psr
+unit-test-coverage: export COVERAGE_EXCLUSIONS ?= ${SOURCE_EXCLUSIONS}
 .PHONY: unit-test-coverage
 unit-test-coverage:  ## run unit tests with coverage
 	${SCRIPT_DIR}/coverage.sh html
@@ -112,7 +114,7 @@ unit-test-coverage-ratcheting:  ## run unit tests with coverage ratcheting
 
 .PHONY: unit-test
 unit-test:  ## run all unit tests in project
-	go test $$(go list ./... | grep -Ev "/tests/e2e|/tools/psr")
+	go test $$(go list ./... | grep -Ev "${SOURCE_EXCLUSIONS}")
 
 #
 #  Compliance check targets
