+++ platform-operator/experimental/controllers/platformctrl/modlifecycle/delegates/interface.go
